[WB-4612] Disable login in disabled mode (#1842)
